apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
python-multipart==0.0.12
pydantic-settings==2.6.1
email-validator==2.2.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
dramatiq[redis]==1.17.1
redis==5.2.0
boto3==1.35.70
pgvector==0.3.6
numpy==2.1.3
scikit-learn==1.5.2
vaderSentiment==3.3.2
pypdf==5.1.0
python-docx==1.1.2
pillow==11.0.0
pytesseract==0.3.13
pandas==2.2.3
orjson==3.10.11
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-exporter-otlp==1.28.2
